Terminated the SessionManager implementation, misses only the docu
