+++ library/src/commonMain/kotlin/com/tecknobit/equinoxcompose/helpers/session/ConnectionChecker.kt
@@ -1,0 +1,7 @@
+package com.tecknobit.equinoxcompose.helpers.session
+
+import androidx.compose.runtime.MutableState
+
+expect fun checkInternetConnection(
+    noInternetConnectionState: MutableState<Boolean>
+)
